pyproject.toml
[tool.poetry]
name = "cw-7-habittracker"
version = "0.1.0"
description = ""
authors = ["Bloodimirg <bloodimirg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
psycopg2-binary = "^2.9.9"
black = "^24.10.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
